Added a check to see if loggers and filters already exist
Since logging.getLogger returns an already existing logger instance
if the name is the same, multiple calls to `logger.getLogger` would
result in multiple handlers and filters being added. This would in
turn result in multiple log lines being printed, 1 for each handler.

This commit ensures that the logger instance returned only has 1
handler or filter attached. A better optimisation in the future
would be to exit out of the function immediately if the handler is
already set when the logger is returned.